WIP: restructured directories and added pyproject.toml file to build online_training package
