+++ .gitignore
@@ -2,4 +2,6 @@
 __pycache__/
 *.py[cod]
 
+*.egg-info
+
