+++ run_scripts/online_training/local/conf/train_config_gnn_small.yaml
