+++ run_scripts/online_training/local/run_ssim_mlp_debug.sh
@@ -2,21 +2,21 @@
 
 # Set executables
 BASE_DIR=/Users/rbalin/Documents/Research/ALCF/SimAI-Bench/SimAI-Bench
-DRIVER=$BASE_DIR/online_training/src/backends/smartsim/ssim_driver.py
-SIM_EXE=$BASE_DIR/online_training/src/data_producers/smartredis/load_data.py
-ML_EXE=$BASE_DIR/online_training/src/train/main.py
+DRIVER=$BASE_DIR/src/online_training/drivers/ssim_driver.py
+SIM_EXE=$BASE_DIR/src/online_training/data_producers/sim.py
+ML_EXE=$BASE_DIR/src/online_training/train/train.py
 TRAIN_CONFIG_PATH=$PWD/conf
-TRAIN_CONFIG_NAME="train_config_gnn_debug"
+TRAIN_CONFIG_NAME="train_config_mlp_debug"
 
 # Set up run
-SIM_RANKS=2
+SIM_RANKS=4
 ML_RANKS=2
 echo Number of simulation ranks: $SIM_RANKS
 echo Number of ML ranks: $ML_RANKS
 echo
 
 # Run
-SIM_ARGS="--model\=gnn --problem_size\=debug --db_launch\=colocated --ppn\=${SIM_RANKS} --tolerance\=0.002"
+SIM_ARGS="--model\=mlp --problem_size\=debug --db_launch\=colocated --ppn\=${SIM_RANKS} --tolerance\=0.002"
 python $DRIVER \
     database.network_interface=lo database.launcher=local \
     sim.executable=$SIM_EXE sim.arguments="${SIM_ARGS}" \
