+++ run_scripts/online_training/polaris/run_ssim_mlp_debug.sh
