+++ run_scripts/online_training/polaris/ssim_config.yaml
