+++ src/online_training/drivers/conf/ssim_config.yaml
