+++ src/online_training/train/gnn/gnn.py
