+++ src/online_training/train/gnn/graph_plotting.py
