+++ src/online_training/train/gnn/model.py
